test: move clienttest package and folder inside the client folder
